+++ examples/05-instancing/instancing.cpp
@@ -118,7 +118,7 @@ class ExampleInstancing : public entry::AppI
 		if (!entry::processEvents(m_width, m_height, m_debug, m_reset) )
 		{
 			// Set view 0 default viewport.
-			bgfx::setViewRect(0, 0, 0, m_width, m_height);
+			bgfx::setViewRect(0, 0, 0, uint16_t(m_width), uint16_t(m_height) );
 
 			// This dummy draw call is here to make sure that view 0 is cleared
 			// if no other draw calls are submitted to view 0.
@@ -178,7 +178,7 @@ class ExampleInstancing : public entry::AppI
 					bgfx::setViewTransform(0, view, proj);
 
 					// Set view 0 default viewport.
-					bgfx::setViewRect(0, 0, 0, m_width, m_height);
+					bgfx::setViewRect(0, 0, 0, uint16_t(m_width), uint16_t(m_height) );
 				}
 
 				const uint16_t instanceStride = 80;
